# Swagger/ucloud-1.0.0-oas3-swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0-oas3
  title: uCloud API collection
  description: uCloud is a collection of image APIs.
  contact:
    email: help@bps-eng.co.ir
  license:
    name: MIT
    url: 'http://www.opensource.org/licenses/MIT'
security:
  - implicit:
      - 'read:face'
      - 'write:face'
...
paths:
...
  /image/barcode:
    post:
      tags:
        - image
      summary: Detect barcodes
      description: Detect and read barcodes found in the image.
      operationId: barcode
      parameters:
        - name: code_type
          in: query
          description: Specify the barcode type.
          required: true
          schema:
            type: string
            enum:
              - auto
              - 2/5 Industrial
              - 2/5 Interleaved
              - Codabar
              - Code 128
              - Code 39
              - Code 93
              - EAN-13 Add-On 2
              - EAN-13 Add-On 5
              - EAN-13
              - EAN-8 Add-On 2
              - EAN-8 Add-On 5
              - EAN-8
              - GS1 DataBar Expanded Stacked
              - GS1 DataBar Expanded
              - GS1 DataBar Limited
              - GS1 DataBar Omnidir
              - GS1 DataBar Stacked Omnidir
              - GS1 DataBar Stacked
              - GS1 DataBar Truncated
              - GS1-128
              - MSI
              - PharmaCode
              - UPC-A Add-On 2
              - UPC-A Add-On 5
              - UPC-A
              - UPC-E Add-On 2
              - UPC-E Add-On 5
              - UPC-E
        - name: dark_on_light
          in: query
          description: Specify the barcode polarity.
          required: false
          schema:
            type: boolean
            default: true
        - name: barcode_height_min
          in: query
          description: Specify minimum barcode height.
          required: false
          schema:
            type: integer
            minimum: 8
            default: -1
        - name: barcode_width_min
          in: query
          description: Specify minimum barcode width.
          required: false
          schema:
            type: integer
            default: -1
        - name: element_size_min
          in: query
          description: Specify minimum element size.
          required: false
          schema:
            type: number
            format: float
            minimum: 0.6
            maximum: 64
            default: 2
        - name: element_size_max
          in: query
          description: Specify maximum element size.
          required: false
          schema:
            type: number
            format: float
            minimum: 1
            maximum: 256
            default: 8
        - name: min_code_length
          in: query
          description: Specify expected minimum code length.
          required: false
          schema:
            type: number
            format: integer
            minimum: 0
            default: 0
        - name: min_contrast
          in: query
          description: Specify minimum barcode contrast.
          required: false
          schema:
            type: number
            format: integer
            minimum: 0
            default: 0
        - name: min_identical_scanlines
          in: query
          description: Specify minimum number of identical scanlines.
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: orientation
          in: query
          description: Specify expected barcode orientation.
          required: false
          schema:
            type: number
            format: float
            minimum: -180
            maximum: 180
            default: 0
        - name: orientation_tolerance
          in: query
          description: Specify barcode orientation tolerance.
          required: false
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 90
            default: 90
      requestBody:
        $ref: '#/components/requestBodies/Image'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Barcode'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Barcode'
      security:
        - implicit:
            - 'read:face'
  /image/datacode:
    post:
      tags:
        - image
      summary: Detect datacodes
      description: Detect and read datacodes (2D barcodes) found in the image.
      operationId: datacode
      parameters:
        - name: code_type
          in: query
          description: Specify the barcode type.
          required: true
          schema:
            type: string
            enum:
              - Aztec Code
              - Data Matrix ECC 200
              - PDF417
              - QR Code
              - Micro QR Code
              - GS1 Aztec Code
              - GS1 DataMatrix
              - GS1 QR Code
        - name: recognition
          in: query
          description: Specify the recognition parameters.
          required: false
          schema:
            type: string
            enum:
              - standard
              - enhanced
              - maximum
            default: standard
        - name: model_type_aztec
          in: query
          description: Specify Aztec model type.
          required: false
          schema:
            type: array
            items:
              type: string
              enum:
                - compact
                - rune
                - full_range
            default: 'compact,full_range'
        - name: symbol_size_min_aztec
          in: query
          description: Specify minimum symbol size for Aztec model.
          required: false
          schema:
            type: number
            format: integer
            minimum: 11
            maximum: 151
            default: 11
        - name: symbol_size_max_aztec
          in: query
          description: Specify maximum symbol size for Aztec model.
          required: false
          schema:
            type: number
            format: integer
            minimum: 11
            maximum: 151
            default: 151
        - name: symbol_cols_min_ecc
          in: query
          description: Specify minimum symbol columns for ECC 200.
          required: false
          schema:
            type: integer
            minimum: 10
            maximum: 144
            default: 10
        - name: symbol_cols_max_ecc
          in: query
          description: Specify maximum symbol columns for ECC 200.
          required: false
          schema:
            type: integer
            minimum: 10
            maximum: 144
            default: 144
        - name: symbol_rows_min_ecc
          in: query
          description: Specify minimum symbol rows for ECC 200.
          required: false
          schema:
            type: integer
            minimum: 8
            maximum: 144
            default: 8
        - name: symbol_rows_max_ecc
          in: query
          description: Specify maximum symbol rows for ECC 200.
          required: false
          schema:
            type: integer
            minimum: 8
            maximum: 144
            default: 144
        - name: model_type_qr
          in: query
          description: Specify QR model type. 0 means both types 1 & 2.
          required: false
          schema:
            type: integer
            enum:
              - 0
              - 1
              - 2
            default: 0
        - name: symbol_size_min_qr
          in: query
          description: Specify minimum symbol size for QR model.
          required: false
          schema:
            type: number
            format: integer
            minimum: 21
            maximum: 177
            default: 21
        - name: symbol_size_max_qr
          in: query
          description: Specify maximum symbol size for QR model.
          required: false
          schema:
            type: number
            format: integer
            minimum: 21
            maximum: 177
            default: 177
        - name: symbol_size_min_uqr
          in: query
          description: Specify minimum symbol size for micro QR models.
          required: false
          schema:
            type: number
            format: integer
            minimum: 11
            maximum: 17
            default: 11
        - name: symbol_size_max_uqr
          in: query
          description: Specify maximum symbol size for micro QR models.
          required: false
          schema:
            type: number
            format: integer
            minimum: 11
            maximum: 17
            default: 17
        - name: symbol_cols_min_pdf
          in: query
          description: Specify minimum symbol columns for PDF-417 models.
          required: false
          schema:
            type: number
            format: integer
            minimum: 1
            maximum: 30
            default: 1
        - name: symbol_cols_max_pdf
          in: query
          description: Specify maximum symbol columns PDF-417 models.
          required: false
          schema:
            type: number
            format: integer
            minimum: 1
            maximum: 30
            default: 20
        - name: symbol_rows_min_pdf
          in: query
          description: Specify minimum symbol rows for PDF-417 models.
          required: false
          schema:
            type: number
            format: integer
            minimum: 3
            maximum: 90
            default: 5
        - name: symbol_rows_max_pdf
          in: query
          description: Specify maximum symbol rows PDF-417 models.
          required: false
          schema:
            type: number
            format: integer
            minimum: 3
            maximum: 90
            default: 45
        - name: contrast_min
          in: query
          description: Specify minimum datacode contrast.
          required: false
          schema:
            type: number
            format: integer
            minimum: 1
            default: 30
        - name: contrast_tolerance
          in: query
          description: Specify datacode contrast tolerance.
          required: false
          schema:
            type: string
            enum:
              - low
              - high
              - any
        - name: dark_on_light
          in: query
          description: Specify the datacode polarity.
          required: false
          schema:
            type: boolean
            default: true
        - name: module_gap_min
          in: query
          description: Specify how big minimum gaps in the symbol are.
          required: false
          schema:
            type: string
            enum:
              - 'no'
              - big
              - small
            default: 'no'
        - name: module_gap_max
          in: query
          description: Specify how big maximum gaps in the symbol are.
          required: false
          schema:
            type: string
            enum:
              - 'no'
              - big
              - small
        - name: slant_max
          in: query
          description: Specify maximum datacode slant from ideal right angle in degrees.
          required: false
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 30
            default: 10
      requestBody:
        $ref: '#/components/requestBodies/Image'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Barcode'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Barcode'
      security:
        - implicit:
            - 'read:face'
...
servers:
  - url: 'https://virtserver.swaggerhub.com/sa-mustafa/ucloud/v1'
  - url: 'http://virtserver.swaggerhub.com/sa-mustafa/ucloud/v1'
components:
  requestBodies:
    Image:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Image'
        application/xml:
          schema:
            $ref: '#/components/schemas/Image'
      description: Image data for further processing
      required: true
...
  securitySchemes:
    implicit:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
          scopes:
            'write:face': allows modifying resources
            'read:face': allows reading resources
    api_key:
      type: apiKey
      name: api_key
      in: header
  schemas:
    Barcode:
      type: object
      properties:
        location:
          type: string
        data:
          type: string
          description: barcode data payload
        type:
          type: string
      xml:
        name: Barcode
...
    Image:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tag:
          type: string
        source:
          type: string
          description: Image URL or content in base64
        type:
          description: Specify the image source type.
          type: string
          enum:
            - url
            - base64
      xml:
        name: Image
...